requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==3.0.1
pinecone==5.0.1
pillow==10.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
aiohttp==9.1
redis==5.0.1
torch==2.2.2
numpy==1.26.2
huggingface-hub==0.20.0
transformers==4.40.2
openai==1.54.0
httpx==0.27.0
httpcore==1.0.5
google-search-results==2.4.2
beautifulsoup4==4.12.3
